+++ src/Control/Monad/Free/Church.hs
@@ -1,11 +1,9 @@
 {-# LANGUAGE BangPatterns #-}
-{-# LANGUAGE CPP #-}
 {-# LANGUAGE Rank2Types #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE Safe #-}
-#include "free-common.h"
 
 -----------------------------------------------------------------------------
 -- |
@@ -48,9 +46,9 @@
 --
 -- This is based on the \"Free Monads for Less\" series of articles by Edward Kmett:
 --
--- * <http://comonad.com/reader/2011/free-monads-for-less/   Free monads for less — Part 1>
+-- * <https://ekmett.github.io/reader/2011/free-monads-for-less/   Free monads for less — Part 1>
 --
--- * <http://comonad.com/reader/2011/free-monads-for-less-2/ Free monads for less — Part 2>
+-- * <https://ekmett.github.io/reader/2011/free-monads-for-less-2/ Free monads for less — Part 2>
 ----------------------------------------------------------------------------
 module Control.Monad.Free.Church
   ( F(..)
@@ -87,7 +85,7 @@ import Prelude hiding (foldr)
 --
 -- It is /asymptotically/ more efficient to use ('>>=') for 'F' than it is to ('>>=') with 'Free'.
 --
--- <http://comonad.com/reader/2011/free-monads-for-less-2/>
+-- <https://ekmett.github.io/reader/2011/free-monads-for-less-2/>
 newtype F f a = F { runF :: forall r. (a -> r) -> (f r -> r) -> r }
 
 -- | Tear down a 'Free' 'Monad' using iteration.
@@ -132,10 +130,8 @@ instance Foldable f => Foldable (F f) where
     foldr f r xs = runF xs f (foldr (.) id) r
     {-# INLINE foldr #-}
 
-#if MIN_VERSION_base(4,6,0)
     foldl' f z xs = runF xs (\a !r -> f r a) (flip $ foldl' $ \r g -> g r) z
     {-# INLINE foldl' #-}
-#endif
 
 instance Traversable f => Traversable (F f) where
     traverse f m = runF m (fmap return . f) (fmap wrap . sequenceA)
@@ -208,9 +204,9 @@ toF xs = F (\kp kf -> go kp kf xs) where
 --
 -- This is based on the \"Free Monads for Less\" series of articles by Edward Kmett:
 --
--- * <http://comonad.com/reader/2011/free-monads-for-less/   Free monads for less — Part 1>
+-- * <https://ekmett.github.io/reader/2011/free-monads-for-less/   Free monads for less — Part 1>
 --
--- * <http://comonad.com/reader/2011/free-monads-for-less-2/ Free monads for less — Part 2>
+-- * <https://ekmett.github.io/reader/2011/free-monads-for-less-2/ Free monads for less — Part 2>
 --
 -- and <http://www.iai.uni-bonn.de/~jv/mpc08.pdf \"Asymptotic Improvement of Computations over Free Monads\"> by Janis Voightländer.
 improve :: Functor f => (forall m. MonadFree f m => m a) -> Free f a